split nltk from utils
